+++ bilheteria/src/modelo/IFilmeDAO.java
@@ -7,9 +7,9 @@ public interface IFilmeDAO {
 
 	public boolean cadastrarFilme(Filme filme);
 
-	public boolean alterarFilme(Filme filme, String nomeFilme, Long idFilme, LocalDateTime horarioFilmeDublado, LocalDateTime horarioFilmeLegendado);
+	public boolean alterarFilme(Filme filme, String nomeFilme, Integer idFilme, LocalDateTime horarioFilmeDublado, LocalDateTime horarioFilmeLegendado);
 
-	public boolean deletarFilme(Filme filme, Long idFilme);
+	public boolean deletarFilme(Filme filme, Integer idFilme);
 
 	public ArrayList<Filme> listaFilmes();
